% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\docType{data}
\name{conda-env}
\alias{conda-env}
\alias{CONDAENV_NAME}
\alias{conda_path}
\alias{env_path}
\alias{configure_matlab}
\alias{configure_conda}
\alias{remove_conda}
\alias{add_packages}
\alias{ensure_rpymat}
\alias{matlab_engine}
\alias{call_matlab}
\title{'Miniconda' environment}
\format{
An object of class \code{character} of length 1.
}
\usage{
CONDAENV_NAME

conda_path()

env_path()

configure_matlab(matlab, python_ver = "auto")

configure_conda(
  python_ver = "auto",
  packages = NULL,
  matlab = NULL,
  update = FALSE,
  force = FALSE
)

remove_conda(ask = TRUE)

add_packages(packages = NULL, python_ver = "auto", ...)

ensure_rpymat()

matlab_engine()

call_matlab(
  fun,
  ...,
  .options = getOption("rpymat.matlab_opt", "-nodesktop -nojvm"),
  .debug = getOption("rpymat.debug", FALSE)
)
}
\arguments{
\item{matlab}{'Matlab' path to add to the configuration path;
see 'Details'}

\item{python_ver}{'Python' version to use; see 'Configuration'}

\item{packages}{additional 'Python' or 'conda' packages to install}

\item{update}{whether to update 'conda'; default is false}

\item{force}{whether to force install the 'Miniconda' even a previous
version exists; default is false. Setting \code{false=TRUE} rarely
works. Please see 'Configuration'.}

\item{ask}{whether to ask for user's agreement to remove the repository.
This parameter should be true if your functions depend on
\code{remove_conda} (see 'CRAN Repository Policy'). This argument might
be removed and force to be interactive in the future.}

\item{...}{for \code{add_packages}, these are additional parameters
passing to \code{\link[reticulate]{conda_install}}; for
\code{call_matlab}, \code{...} are the parameters passing to \code{fun}}

\item{fun}{'Matlab' function name, character (experimental)}

\item{.options}{'Matlab' compiler options}

\item{.debug}{whether to enable debug mode}
}
\value{
None
}
\description{
These functions/variables are used to configure
'Miniconda' environment.
}
\section{Background & Objectives}{

Package 'reticulate' provides sophisticated tool-sets that
allow us to call 'Python' functions within \code{R}. However, the
installation of 'Miniconda' and 'Python' can be tricky on many
platforms, for example, the 'M1' chip, or some other 'ARM' machines.
The package \code{rpymat} provides easier approach to configure on these
machines with totally isolated environments. Any modifications to this
environment will not affect your other set ups.

Since 2014, 'Matlab' has introduced its official compiler for 'Python'.
The package \code{rpymat} provides a simple approach to link the
compiler, provided that you have proper versions of 'Matlab' installed.
\href{https://www.mathworks.com/content/dam/mathworks/mathworks-dot-com/support/sysreq/files/python-compatibility.pdf}{Here} is a list of
'Matlab' versions with official compilers and their corresponding
'Python' versions.
}

\section{Configuration}{

If 'Matlan' compiler is not to be installed, In most of the cases,
function \code{configure_conda} with default arguments automatically
downloads the latest 'Miniconda' and configures the latest 'Python'.
If any other versions of 'Miniconda' is ought to be installed,
please set options \code{"reticulate.miniconda.url"} to change the
source location.

On Apple's 'M1' machine, there is no official native 'Miniconda' as of
2021-11-01. The default 'conda' will be the \href{https://github.com/conda-forge/miniforge/releases/tag/4.9.2-7}{'Mambaforge'} compiled by
the 'conda-forge' team. This version will subject to change in the near
future once the official support is released. Sometimes the 'Mambaforge'
is too advanced and not all the dependencies are properly compiled.
In such cases, you will receive errors like "Package 'xxx' conflicts
for". Try a lower version of \code{python_ver} (such as "3.9").

If 'Matlab' is to be installed, please specify the 'Matlab' path when
running \code{configure_conda}. If the environment has been setup,
\code{configure_matlab} can link the 'Matlab' compilers without
removing the existing environment. For 'ARM' users, unfortunately,
there will be no 'Matlab' support as the compilers are written for
the 'Intel' chips.
}

\section{Initialization}{

Once 'conda' and 'Python' environment has been installed, make sure
you run \code{ensure_rpymat()} before running any 'Python' code. This
function will make sure correct compiler is linked to your current
\code{R} session.
}

\examples{

# The script will interactively install 'conda' to `R_user_dir`
\dontrun{

# Install conda and python 3.9

configure_conda(python_ver = '3.9')


# Add packages h5py, pandas, jupyter

add_packages(c('h5py', 'pandas', 'jupyter'))

# Add pip packages

add_packages("itk", pip = TRUE)

# Initialize the isolated environment

ensure_rpymat()


# Remove the environment

remove_conda()

}

}
\keyword{datasets}
