% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{custom_env_support}
\alias{custom_env_support}
\title{Whether supports customized \code{'conda'} environment}
\usage{
custom_env_support()
}
\value{
true or false on whether multiple customized environment is
supported
}
\description{
Whether supports customized \code{'conda'} environment
}
