% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_py.R
\name{repl_python}
\alias{repl_python}
\title{Enable interactive 'python' from R}
\usage{
repl_python(..., env_name = NA)
}
\arguments{
\item{...}{passed to \code{\link[reticulate]{repl_python}}
in \code{'reticulate'} package}

\item{env_name}{environment name to activate, if not default. This argument
is ignored if any other environment is activated; see
\code{\link{ensure_rpymat}}.}
}
\value{
See \code{\link[reticulate]{repl_python}}
}
\description{
Allows users to type 'python' command from R
console for quick code evaluation or debugging.
}
