% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{rpymat-python-main}
\alias{rpymat-python-main}
\alias{py}
\title{Get 'Python' main process environment}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
py
}
\value{
The 'Python' main process as a module
}
\description{
\code{py} automatically converts 'Python' objects to R objects.
\code{\link{import_main}} does not convert by default; see 'Examples' for details.
}
\examples{

py_no_convert <- rpymat::import_main(convert = FALSE)

py$a <- matrix(seq_len(16), 4)

py_no_convert$a

py$a

}
\keyword{datasets}
