% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_py.R
\name{run_script}
\alias{run_script}
\title{Run 'Python' script}
\usage{
run_script(x, work_dir = NULL, local = FALSE, convert = FALSE)
}
\arguments{
\item{x}{script path}

\item{work_dir}{working directory of the script}

\item{local, convert}{passed to \code{\link[reticulate]{py_run_file}}}
}
\value{
The values returned by \code{\link[reticulate]{py_run_file}}
}
\description{
A wrapper of \code{\link[reticulate]{py_run_file}},
but with \code{rpymat} enabled
}
\examples{

\dontrun{

# Please configure conda environment first

x <- tempfile()
writeLines(c(
  "import re",
  "zipcode = re.findall(r'[0-9]{5,6}', r.address)"
), con = x)

address <- '2341 Main St., 72381'
rpymat::run_script(x)

py$zipcode

}

}
