% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_py.R
\name{repl_python}
\alias{repl_python}
\title{Enable interactive 'python' from R}
\usage{
repl_python(...)
}
\arguments{
\item{...}{passed to \code{\link[reticulate]{repl_python}}
in \code{'reticulate'} package}
}
\value{
See \code{\link[reticulate]{repl_python}}
}
\description{
Allows users to type 'python' command from R
console for quick code evaluation or debugging.
}
